Optimize hash table size and use GCC likely extention
Toyed around with the hash table size. Also used the likely extension
since insertion only happens once, but after that you are likely to be
doing string checks. This yielded a good 4% speedup, but it's biased by
the fact that the actual number of valid station names is in the 400s
not 10000. Current mean: 463 ms