Optimize hash table for short strings
This adds a new StationName that stores short strings as aligned 32 byte
strings. This allows us to use optimal load/comparison instructions to
speed up comparisons in the hash table. It appears that the cost of the
copy is much lower than the cost of the comparison - probably because
insertions are much much less frequent than comparisons.

There are some known gaps - the code isn't really well tested on longer
strings, and there is the possibility that the code accesses out of
bounds data. Since mmap always creates memory-aligned addresses, this
should only happen if the end of the underlying data is very close to
the end of the page. There should be some way around this, but I haven't
gotten around to it yet.

This created about a 25% speedup, bringing the runtime down to just
under 300ms.